feat: product

+++ src/domain/user/init.go
@@ -4,7 +4,7 @@ import "context"
 
 type Entity interface {
 	GetListUser(ctx context.Context, offset, limit int, search, sort string) (res []User, count int, err error)
-	GetUserById(ctx context.Context, id int) (res User, err error)
+	GetUserById(ctx context.Context, id int64) (res User, err error)
 	Login(ctx context.Context, b Login) (res UserLogin, err error)
 	Create(ctx context.Context, b SetUser) (err error)
 	Update(ctx context.Context, b SetUser, id int) (err error)
@@ -12,7 +12,7 @@ type Entity interface {
 }
 type Repo interface {
 	GetListUser(ctx context.Context, offset, limit int, search, sort string) (res []User, count int, err error)
-	GetUserById(ctx context.Context, id int) (res User, err error)
+	GetUserById(ctx context.Context, id int64) (res User, err error)
 	Login(ctx context.Context, b Login) (res UserLogin, err error)
 	Create(ctx context.Context, b SetUser) (err error)
 	Update(ctx context.Context, b SetUser, id int) (err error)
